requirements.txt
torch
torchvision
cupy-cuda12x
transformers==4.46.2
controlnet-aux==0.0.7
imageio
imageio[ffmpeg]
safetensors
einops
sentencepiece
protobuf
modelscope

peft
# lightning
gradio
decord
ftfy
pandas
pillow
numpy
